[chore] refactoring 후 class package 명시 수정
