+++ src/main/java/com/soolsul/soolsulserver/domain/common/BaseEntity.java
@@ -1,0 +1,4 @@
+package com.soolsul.soolsulserver.domain.common;
+
+public class BaseEntity {
+}
