+++ src/main/java/com/soolsul/soolsulserver/domain/curation/CurationPicture.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.domain;
+package com.soolsul.soolsulserver.domain.curation;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
