+++ src/main/java/com/soolsul/soolsulserver/domain/region/Region.java
@@ -1,4 +1,4 @@
-package com.soolsul.soolsulserver.domain;
+package com.soolsul.soolsulserver.domain.region;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
